chore: improve loading, cleanup
